More clang-tidy fixes

+++ amr-wind/equation_systems/icns/source_terms/ABLWrfForcingMom.cpp
@@ -365,7 +365,7 @@ void ABLWrfForcingMom::operator()(
     const auto& problo = m_mesh.Geom(lev).ProbLoArray();
     const auto& dx = m_mesh.Geom(lev).CellSizeArray();
 
-    const int nh_max = m_velAvg_ht.size() - 2;
+    const int nh_max = (int)m_velAvg_ht.size() - 2;
     const int lp1 = lev + 1;
     const amrex::Real* vheights = m_wrf_ht.data();
     const amrex::Real* u_error_val = m_error_wrf_avg_U.data();
